COMMERCE-7143 batch-planner-web - move source to appropriate js
